requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pillow==10.0.0
matplotlib==3.7.2
tqdm==4.65.0 